requirements.txt
Flask==2.3.3
tensorflow-cpu==2.14.0
numpy==1.24.3
Werkzeug==2.3.7
gunicorn==21.2.0
Pillow==10.0.0